maty - attach detach interface move to serverApi
