+++ apis/openstack-nova/src/main/java/org/jclouds/openstack/nova/v2_0/features/InterfaceApi.java
@@ -37,16 +37,5 @@ import javax.ws.rs.core.MediaType;
 @Path("/servers")
 public interface InterfaceApi {
 
-   @POST
-   @Consumes(MediaType.APPLICATION_JSON)
-   @SelectJson("interfaceAttachment")
-   @Path("/{server_id}/os-interface")
-   @RequestFilters(AuthenticateRequest.class)
-   InterfaceAttachment attachInterface(@PathParam("server_id") String serverId, AttachInterfaceOptions... options);
 
-   @DELETE
-   @Consumes(MediaType.APPLICATION_JSON)
-   @Path("/{server_id}/os-interface/{port_id}")
-   @RequestFilters(AuthenticateRequest.class)
-   void detachInterface(@PathParam("server_id") String serverId, @PathParam("port_id") String portId);
 }
